pyproject.toml
[project]
name = "legal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.1",
    "langgraph>=1.0.3",
    "pydantic>=2.12.4",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.40.0",
    "uvicorn>=0.38.0",
]
